+++ src/test/java/com/cedarsoftware/io/CustomJsonTest.java
@@ -79,13 +79,11 @@ class CustomJsonTest
 	{
 		ReadOptionsBuilder readOptions = new ReadOptionsBuilder()
 				.addClassFactory(Person.class, new PersonFactory())
-				.aliasTypeName(Person.class, "Person")
-				.withExtendedAliases();
+				.aliasTypeName(Person.class, "Person");
 
 		WriteOptionsBuilder writeOptions = new WriteOptionsBuilder()
 				.addCustomWrittenClass(Person.class, new PersonWriter())
-				.aliasTypeName(Person.class, "Person")
-				.withExtendedAliases();
+				.aliasTypeName(Person.class, "Person");
 
 		Person p1 = createPersonJohn();
 		String json = JsonIo.toJson(p1, writeOptions.build());