+++ changelog.md
@@ -1,4 +1,7 @@
 ### Revision History
+* 4.24.0
+  * All aliases have been moved to aliases.txt in the resouces folder. It is a very complete list of class names to alias names. If you want less aliases (or more), use the `addPermanentAlias()` APIs on `ReadOptionsBuilder` and `WriteOptionsBuilder.` If you do not want a particular alias output, use `WriteOptionsBuilder.removeAliasedClassName(wildcardPattern).`
+  * The "extendedAliases" option has been removed from Read/Write options builders. By default, as many aliases are enabled as possible, and you can use the removePermanentXXX API to reduce them, or place your own version of `aliases.txt` in the classpath ahead of the one in `json-io.jar.`
 * 4.23.0
   * `Collections.unmodifiableXXX()` instances when serialized, restore back to unmodifiable instances.
   * `ImmutableList` and `ImmutableSet` restore back unmodifiable instances.